Product CRUD , sorting , pagination , filtering

+++ controllers/categoryController.js
@@ -1,4 +1,5 @@
 const Category = require("../models/categoryModel");
+const Products = require("../models/productModel");
 
 const categoryController = {
   getCategories: async (req, res) => {
@@ -28,10 +29,11 @@ const categoryController = {
   },
   deleteCategory: async (req, res) => {
     try {
-      // const products = await Products.findOne({category: req.params.id})
-      // if(products) return res.status(400).json({
-      //     msg: "Please delete all products with a relationship."
-      // })
+      const products = await Products.findOne({ category: req.params.id });
+      if (products)
+        return res.status(400).json({
+          msg: "Please delete all products with a relationship.",
+        });
 
       await Category.findByIdAndDelete(req.params.id);
       res.json({ msg: "Deleted a Category" });
